+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ## Fixes
 
 - Resolved the `Microsoft.Extensions.FileSystemGlobbing.dll` dependency conflict ([#1253](https://github.com/getsentry/sentry-unity/pull/1253))
+### Fixes
+
+- Sentry CLI integration when exporting an Android project ([#1242](https://github.com/getsentry/sentry-unity/pull/1242))
 
 ### Dependencies